requirements.txt
# PainCare AI Model Dependencies

# Core ML and AI libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# XAI (Explainable AI) libraries
lime>=0.2.0
shap>=0.42.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# API and Web Framework
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
python-jose>=3.3.0
python-multipart>=0.0.6

# Firebase integration
firebase-admin>=6.2.0

# Data processing and utilities
scipy>=1.11.0
joblib>=1.3.0
python-dateutil>=2.8.0
pytz>=2023.3

# Security and authentication
cryptography>=41.0.0
bcrypt>=4.0.0
passlib>=1.7.0

# HTTP and requests
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0
aiofiles>=23.2.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Jupyter notebooks for experimentation
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Monitoring and logging
loguru>=0.7.0

# Development tools
black>=23.7.0
flake8>=6.0.0
